feat: add controllers revision

+++ Customer.Datalayer/src/Customer.Datalayer.WebApi/Controllers/AddressController.cs
@@ -1,5 +1,4 @@
 ﻿using Customer.Datalayer.BusinessEntities;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Customer.Datalayer.WebApi.Controllers
@@ -16,32 +15,40 @@ namespace Customer.Datalayer.WebApi.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<List<Addresses>>> Get()
+        public async Task<ActionResult<List<Addresses>>> Get() // READ ALL
         {
             return Ok(await dbContext.Addresses.ToListAsync());
         }
 
         [HttpPost]
-        public async Task<ActionResult<List<Addresses>>> AddCustomer(Addresses address)
+        public async Task<ActionResult<List<Addresses>>> AddCustomer(Addresses address) // CREATE
         {
+            if (address == null) return BadRequest("Address not found.");
+
             dbContext.Addresses.Add(address);
             await dbContext.SaveChangesAsync();
+
             return Ok(await dbContext.Addresses.ToListAsync());
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Addresses>> Get(int id)
+        public async Task<ActionResult<Addresses>> Get(int id) // READ
         {
+            if (id == 0) return BadRequest("This ID will not be found.");
+
             var address = await dbContext.Addresses.FindAsync(id);
             if (address == null) return BadRequest("Address not found");
+
             return Ok(address);
         }
 
         [HttpPut]
-        public async Task<ActionResult<List<Addresses>>> UpdateCustomer(Addresses address)
+        public async Task<ActionResult<List<Addresses>>> UpdateCustomer(Addresses address) // UPDATE
         {
+            if (address == null) return BadRequest("Address not found.");
+
             var dbAddress = await dbContext.Addresses.FindAsync(address.AddressId);
-            if (dbAddress == null) return BadRequest("Customer not found");
+            if (dbAddress == null) return BadRequest("Customer not found.");
 
             dbAddress.CustomerId = address.CustomerId;
             dbAddress.AddressLine = address.AddressLine;
@@ -58,8 +65,10 @@ namespace Customer.Datalayer.WebApi.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<List<Addresses>>> Delete(int id)
+        public async Task<ActionResult<List<Addresses>>> Delete(int id) // DELETE
         {
+            if (id == 0) return BadRequest("This ID will not be found.");
+
             var address = await dbContext.Addresses.FindAsync(id);
             if (address == null) return BadRequest("Address not found");
             dbContext.Addresses.Remove(address);
